extract Electron logic to a hook
