+++ Composer/packages/client/src/pages/design/DesignPage.tsx
@@ -44,6 +44,7 @@ import {
   actionsSeedState,
   userSettingsState,
 } from '../../recoilModel';
+import { useElectronFeatures } from '../../hooks/useElectronFeatures';
 
 import {
   breadcrumbClass,
@@ -226,7 +227,7 @@ const DesignPage: React.FC<RouteComponentProps<{ dialogId: string; projectId: st
   const { actionSelected, showDisableBtn, showEnableBtn } = useMemo(() => {
     const actionSelected = Array.isArray(visualEditorSelection) && visualEditorSelection.length > 0;
     if (!actionSelected) {
-      return {};
+      return { actionSelected: false, showDisableBtn: false, showEnableBtn: false };
     }
     const selectedActions = visualEditorSelection.map((id) => get(currentDialog?.content, id));
     const showDisableBtn = selectedActions.some((x) => get(x, 'disabled') !== true);
@@ -234,37 +235,9 @@ const DesignPage: React.FC<RouteComponentProps<{ dialogId: string; projectId: st
     return { actionSelected, showDisableBtn, showEnableBtn };
   }, [visualEditorSelection]);
 
-  // Sync selection state to Electron main process
-  useEffect(() => {
-    if (!window.__IS_ELECTRON__) return;
-    if (!window.ipcRenderer || typeof window.ipcRenderer.send !== 'function') return;
-
-    window.ipcRenderer.send('composer-state-change', { actionSelected });
-  }, [actionSelected]);
+  useElectronFeatures(actionSelected);
 
   const EditorAPI = getEditorAPI();
-  // Subscribe Electron app menu events (copy/cut/del/undo/redo)
-  useEffect(() => {
-    if (!window.__IS_ELECTRON__) return;
-    if (!window.ipcRenderer || typeof window.ipcRenderer.on !== 'function') return;
-
-    window.ipcRenderer.on('electron-menu-clicked', (e, data) => {
-      const label = get(data, 'label', '');
-      switch (label) {
-        case 'undo':
-          return EditorAPI.Editing.Undo();
-        case 'redo':
-          return EditorAPI.Editing.Redo();
-        case 'cut':
-          return EditorAPI.Actions.CutSelection();
-        case 'copy':
-          return EditorAPI.Actions.CopySelection();
-        case 'delete':
-          return EditorAPI.Actions.DeleteSelection();
-      }
-    });
-  }, []);
-
   const toolbarItems: IToolbarItem[] = [
     {
       type: 'dropdown',
